milestone 4

+++ app/src/main/java/com/selflearningcoursecreationapp/ui/create_course/upload_content/UploadContentRepo.kt
@@ -13,7 +13,7 @@ interface UploadContentRepo {
         lectureId: Int,
         file: File,
         uploadType: Int,
-        duration: Int,
+        duration: Long,
     ): Flow<Resource>
 
     suspend fun thumbnailUpload(
